refactor: additional refactorings of package structure and transpose flow graph logic
