+++ bundles/mdpa.gdpr.analysis/src/mdpa/gdpr/analysis/GDPRLegalAssessmentAnalysisBuilder.java
@@ -1,8 +1,8 @@
 package mdpa.gdpr.analysis;
 
 import java.util.Optional;
-import mdpa.gdpr.analysis.core.resource.GDPRResourceProvider;
-import mdpa.gdpr.analysis.core.resource.GDPRURIResourceProvider;
+import mdpa.gdpr.analysis.resource.GDPRResourceProvider;
+import mdpa.gdpr.analysis.resource.GDPRURIResourceProvider;
 import org.apache.log4j.Logger;
 import org.dataflowanalysis.analysis.DataFlowAnalysisBuilder;
 import org.dataflowanalysis.analysis.utils.ResourceUtils;
