+++ bundles/mdpa.gdpr.analysis/src/mdpa/gdpr/analysis/core/ContextDependentAttributeScenario.java
@@ -1,7 +1,7 @@
 package mdpa.gdpr.analysis.core;
 
 import java.util.List;
-import mdpa.gdpr.analysis.UncertaintyUtils;
+import mdpa.gdpr.analysis.utils.UncertaintyUtils;
 import mdpa.gdpr.analysis.dfd.DFDGDPRTransposeFlowGraph;
 import mdpa.gdpr.analysis.dfd.DFDGDPRVertex;
 import mdpa.gdpr.metamodel.contextproperties.Expression;
