+++ bundles/mdpa.gdpr.analysis/src/mdpa/gdpr/analysis/core/TransformationManager.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import mdpa.gdpr.analysis.dfd.DataFlowDiagramAndDataDictionary;
+import mdpa.gdpr.analysis.resource.DataFlowDiagramAndDataDictionary;
 import mdpa.gdpr.dfdconverter.GDPR2DFD;
 import mdpa.gdpr.dfdconverter.tracemodel.tracemodel.NodeTrace;
 import mdpa.gdpr.dfdconverter.tracemodel.tracemodel.TraceModel;
