+++ bundles/mdpa.gdpr.analysis/src/mdpa/gdpr/analysis/dfd/DFDGDPRFlowGraphCollection.java
@@ -7,7 +7,8 @@ import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
 import mdpa.gdpr.analysis.core.ContextDependentAttributeSource;
-import mdpa.gdpr.analysis.core.resource.GDPRResourceProvider;
+import mdpa.gdpr.analysis.resource.DataFlowDiagramAndDataDictionary;
+import mdpa.gdpr.analysis.resource.GDPRResourceProvider;
 import mdpa.gdpr.metamodel.GDPR.AbstractGDPRElement;
 import mdpa.gdpr.metamodel.GDPR.PersonalData;
 import mdpa.gdpr.metamodel.GDPR.Processing;
